less precision loss score log

+++ util/score/score.go
@@ -12,17 +12,17 @@ func sign(x int32) int64 {
 	return 1
 }
 
-func log(x int32) int64 {
-	return int64(math.Pow(math.Log1p(math.Abs(float64(x))), 3))
+func log(x int32) float64 {
+	return math.Pow(math.Log1p(math.Abs(float64(x))), 3)
 }
 
 const (
-	timeSlope int64 = 64
+	timeSlope float64 = 64
 )
 
 // Log is the logarithmic sort
 func Log(ups, downs int32, creationTime int64) int64 {
 	x := ups - downs
 	y := creationTime - time.Now().Unix()
-	return sign(x)*log(x)*timeSlope + y
+	return sign(x)*int64(log(x)*timeSlope) + y
 }
